backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==2.3.7
